% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ink-package.R
\docType{package}
\name{ink-package}
\alias{ink}
\alias{ink-package}
\title{ink: The Modern, High-Performant, Graphic Device for R}
\description{
Provides state-of-the-art quality and performance graphic devices
    based on the 'Blend2D' graphic framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/ink}
  \item Report bugs at \url{https://github.com/thomasp85/ink/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@rstudio.com} (0000-0002-5147-4711)

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
