% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ink_dev.R
\name{ink_bmp}
\alias{ink_bmp}
\title{Draw to a bmp file}
\usage{
ink_bmp(
  filename = "Rplot\%03d.bmp",
  width = 480,
  height = 480,
  units = "px",
  pointsize = 12,
  background = "white",
  res = 72
)
}
\arguments{
\item{filename}{The name of the file. Follows the same semantics as the file
naming in \code{\link[grDevices:png]{grDevices::png()}}, meaning that you can provide a \code{\link[=sprintf]{sprintf()}}
compliant string format to name multiple plots (such as the default value)}

\item{width, height}{The dimensions of the device}

\item{units}{The unit \code{width} and \code{height} is measured in, in either pixels
(\code{'px'}), inches (\code{'in'}), millimeters (\code{'mm'}), or centimeter (\code{'cm'}).}

\item{pointsize}{The default pointsize of the device in pt}

\item{background}{The background colour of the device}

\item{res}{The resolution of the device. This setting will govern how device
dimensions given in inches, centimeters, or millimeters will be converted
to pixels. Further, it will be used to scale text sizes and linewidths}
}
\description{
The BMP (bitmap) format is an image format developed by Microsoft to store
raster images. It is relatively simple and well documented making it a
popular lower bound format. In general, PNG is now prefered over BMP if
possible.
}
\examples{
file <- tempfile(fileext = '.ppm')
ink_bmp(file)
plot(sin, -pi, 2*pi)
dev.off()

}
